refactor: split 9 oversized modules into single-responsibility files (#61)
Split all files exceeding the 300-line threshold, following the extraction
pattern from 437f043. 19 new focused modules created; no re-exports.

- orchestrator.py (2021→566): extract method bodies to 8 modules
  (variant, fields, wiring, inspect, write, validation, postcondition, patch).
  Phase1Orchestrator stays as thin router with wrapper methods.
- smoke_history.py (1173→284): extract stats, report, pipeline modules.
- smoke_batch.py (786→393): extract case resolution + runner modules.
- material_inspector.py (633→271): extract variant inspection module.
- csharp_fields.py (603→316): extract resolution/inheritance module.
- material_asset_inspector.py (577→386): extract property writer module.
- udon_wiring.py (380→269): extract MonoBehaviour field parser module.
- unity_assets.py (324→222): extract path resolution utilities module.
- reporting.py (305→156): extract markdown rendering module.

Tests: 1494 passed (excl. Unity-env-dependent), 0 new failures.
