+++ tests/test_material_inspector.py
@@ -11,11 +11,11 @@ from prefab_sentinel.material_inspector import (
     MaterialSlot,
     RendererMaterials,
     _inspect_base_materials,
-    _parse_material_overrides,
     _parse_renderer_materials,
     format_materials,
     inspect_materials,
 )
+from prefab_sentinel.material_inspector_variant import _parse_material_overrides
 from prefab_sentinel.unity_yaml_parser import YamlBlock
 from tests.yaml_helpers import (
     YAML_HEADER,
@@ -966,7 +966,7 @@ class TestNestedRecursiveTraversal(unittest.TestCase):
 
     def test_depth_limit_stops_recursion(self) -> None:
         """A chain of 12 nested prefabs: renderer at depth 10 IS reachable, depth 11 is NOT."""
-        from prefab_sentinel.material_inspector import _collect_nested_renderers
+        from prefab_sentinel.material_inspector_variant import _collect_nested_renderers
 
         with tempfile.TemporaryDirectory() as tmp:
             tmp_path = Path(tmp)
@@ -1047,7 +1047,7 @@ class TestNestedRecursiveTraversal(unittest.TestCase):
         Currently it only goes 1 level deep. After the fix it should find
         Leaf's renderer through Mid's PrefabInstance.
         """
-        from prefab_sentinel.material_inspector import _collect_nested_renderers
+        from prefab_sentinel.material_inspector_variant import _collect_nested_renderers
 
         with tempfile.TemporaryDirectory() as tmp:
             tmp_path = Path(tmp)
