+++ tests/test_udon_wiring.py
@@ -7,10 +7,12 @@ import unittest
 from prefab_sentinel.contracts import Severity
 from prefab_sentinel.udon_wiring import (
     SKIP_FIELDS,
-    UDON_BEHAVIOUR_GUID,
-    _parse_monobehaviour_fields,
     analyze_wiring,
     extract_monobehaviour_field_names,
+)
+from prefab_sentinel.udon_wiring_parser import (
+    UDON_BEHAVIOUR_GUID,
+    _parse_monobehaviour_fields,
 )
 from prefab_sentinel.unity_yaml_parser import (
     YamlBlock,
